Took meeting 10 notes and did some navigation theory implementation with Esteban.
Implemented functional programming paradigm inside the LoginFragment.kt. We now require an argument when creating an instance of LoginFragment (like is the case of the one given to supportFragmentManager inside UserAccessActivity.kt). This way we do not require an Navigator interface nor it's implementations. We now have more flexibility and less coupling.

Three different ways to implement the afore mentioned can be saw in AccessNavigator.kt