Implemented BottomNavigation UI along with its styles and resources
Decisions on the structure of the styles defined in themes.xml are based in materialUI guidelines.

Structure of BottomNavigation component is as necessary for the correct label displays and colors.

Background shape was particularly challenging. If not solid attribute was defined, it would not be shown well. In order to use the theme styles, I had to set the solid color to ?attr/colorPrimary, which is particularly the value overwritten in the theme style